requirements.txt
aiofiles==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
ffmpeg-python==0.2.0
filelock==3.9.0
future==0.18.3
gunicorn==20.1.0
httptools==0.5.0
huggingface-hub==0.13.1
idna==3.4
llvmlite==0.39.1
more-itertools==9.1.0
multidict==6.0.4
numba==0.56.4
numpy==1.23.5
openai-whisper==20230308
packaging==23.0
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
sanic==22.12.0
Sanic-Cors==2.2.0
sanic-ext==22.12.0
sanic-routing==22.8.0
tokenizers==0.13.2
torch==1.13.1
tqdm==4.65.0
transformers==4.26.1
typing_extensions==4.5.0
ujson==5.7.0
urllib3==1.26.15
uvloop==0.17.0
websockets==10.4
